Move MySQL & SQLite wrappers into the clp namespace.
