+++ components/core/src/clp/SQLiteDB.cpp
@@ -1,10 +1,11 @@
 #include "SQLiteDB.hpp"
 
-#include "Defs.h"
-#include "spdlog_with_specializations.hpp"
+#include "../Defs.h"
+#include "../spdlog_with_specializations.hpp"
 
 using std::string;
 
+namespace clp {
 void SQLiteDB::open(string const& path) {
     auto return_value = sqlite3_open(path.c_str(), &m_db_handle);
     if (SQLITE_OK != return_value) {
@@ -36,3 +37,4 @@ SQLiteDB::prepare_statement(char const* statement, size_t statement_length) {
 
     return {statement, statement_length, m_db_handle};
 }
+}  // namespace clp
